requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pillow>=10.0
opencv-python-headless>=4.9
numpy>=1.26
google-generativeai>=0.5
realesrgan>=0.3
